# api/openapi/user/v1/api.yaml
openapi: 3.0.1
info:
  version: 0.3.0
  title: Meower REST API
  description: API for common CRUD operations.
servers:
  - url: https://meower.com/{basePath}
    description: The production API server.
    variables:
      basePath:
        default: api/v1
security:
  - oAuth:
    - general
paths:
  /users/{userID}:
    parameters:
      - name: userID
        in: path
        required: true
        description: The id of the user.
        schema:
          type: string
    put:
      tags:
        - user
      summary: Update the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated.
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    oAuth:
      type: oauth2
      description: This API uses OAuth2 with the “Authorization Code with Proof Key of Code Exchange (PKCE)” (RFC7636) flow.
      flows:
         authorizationCode:
          authorizationUrl: https://auth.meower.com/oauth/v2/authorize
          tokenUrl: https://auth.meower.com/oauth/v2/token
          refreshUrl: https://auth.meower.com/oauth/v2/token
          scopes:
            general: ""
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          description: The id of the user.
          example: NArmstrong
        name:
          type: string
          description: The name of the user.
          example: Neil Armstrong
        imageUrl:
          type: string
          format: uri
          description: The image url of the user.
        statusText:
          type: string
          description: The status text of the user.
      required:
        - id
        - name
        - imageUrl
        - statusText
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Description of the error.
      required:
        - error