# internal/delivery/http/gen/user/v1/models.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package v1

const (
	OAuthScopes = "oAuth.Scopes"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Error Description of the error.
	Error string `json:"error"`
}

// User defines model for User.
type User struct {
	// Id The id of the user.
	Id string `json:"id"`

	// ImageUrl The image url of the user.
	ImageUrl string `json:"imageUrl"`

	// Name The name of the user.
	Name string `json:"name"`

	// StatusText The status text of the user.
	StatusText string `json:"statusText"`
}

// PutUsersUserIDJSONRequestBody defines body for PutUsersUserID for application/json ContentType.
type PutUsersUserIDJSONRequestBody = User